+++ assets/js/frontend/containers/IdAndPassport/avatarColors.ts
